refactor: src structure
