+++ src/nvshim/compiler/installer.py
@@ -1,4 +1,4 @@
-from installer.__main__ import main
+from nvshim.installer.__main__ import main
 
 
 if __name__ == "__main__":
